[Submit] Infer body from commit message (#366)
This PR allows the CLI to parse the submit message from a multi-line commit.

(new to me) Git itself parses its commit messages with a 'subject' and a 'body'. So far, our commit message logic has just extracted the 'subject' (aka the title) of the commit.

This PR 1) adds support for grabbing the 'body' and 2) hooks that up to our submit logic.

This PR was authored and submitted with this new functionality.